refactor(config): move the cfg reader and writer out of the entry point
PERSO.CPP holds main(), which is why the cfg reader and writer ended up there
and not for any better reason. Move them to CONFIG_FILE.CPP, taking the CONFIG
banner along so it still heads the code it names. PERSO.CPP loses 390 lines.

CONFIG_FILE.CPP is not a module in the sense FOLLOWCAM.CPP is: it owns almost
nothing it touches, because serialising other people's settings is the job. Its
header declares three functions and no state, and the three statics it does own
are the per-run override bookkeeping WriteConfigFile needs to tell a forced
value from a changed one.

So the ownership rule inverts for it, and the camera moves first: FOLLOWCAM.CPP
gains FollowCam_ReadConfig and FollowCam_WriteConfig, and CONFIG_FILE.CPP calls
those instead of naming eleven camera keys itself. A camera setting is now added
in one file rather than two, and FollowCamGroundClearance's clamp sits in the
same translation unit as the occlusion march whose S32 range is the only reason
that clamp exists.

The FollowCamera master toggle is deliberately left behind. The cfg is rewritten
in key order and that toggle sits apart from the camera block, so moving it
would reorder every player's lba2.cfg on the next write.

StoredLanguage was the one piece of the block's private state read from outside:
boot captures the cfg's language before --language may replace it. That becomes
ConfigFile_NoteBootLanguage rather than an exported variable.

Verified three ways against the previous commit: the written lba2.cfg is
byte-identical with its key order intact, --dump-state matches on every line but
the two wall-clock ones, and a camtrace run matches for all 77 frames. 150/150
in the container; the camera fixtures pass.
